pyproject.toml
[project]
name = "lares"
version = "0.1.0"
description = "A stateful AI agent with persistent memory - your household guardian"
readme = "README.md"
license = {text = "PolyForm-Noncommercial-1.0.0"}
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "feedparser>=6.0.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
lares = "lares.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lares"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
